add support for ceil build function
