+++ classes/structure.py
@@ -57,6 +57,9 @@ class Structure(Basic):
             return None
         return len(self.behavior) - 1
 
+    def insertBehavior(self, index: int, element: Protocol):
+        self.behavior.insert(index, element)
+
     def addProtocol(
         self,
         protocol_identifier: str,
@@ -69,7 +72,7 @@ class Structure(Basic):
             if self.parametrs is not None:
                 tmp += self.parametrs
             if parametrs is not None:
-                tmp.elements += parametrs
+                tmp += parametrs
         else:
             tmp = parametrs
         self.behavior.append(Protocol(protocol_identifier, (0, 0), element_type, tmp))
