+++ examples/build_in_functions/ceil/aplan/project.env_descript
@@ -1,0 +1,22 @@
+environment (
+	types : obj (
+			Nil
+	);
+	attributes : obj (Nil);
+	agent_types : obj (
+		CEIL_EXAMPLE : obj (
+			fractional_part:float,
+			integral_part:int,
+			result_ceil:int,
+			input_val:float,
+			result:float
+		),
+		ENVIRONMENT:obj(Nil)
+	);
+	agents : obj (
+		CEIL_EXAMPLE : obj (ceil_example),
+		ENVIRONMENT : obj (env)
+	);
+	axioms : obj (Nil);
+	logic_formula : obj (1)
+);
